Revised backend get_data function for easier testing.
Removed duration checking for the moment, will bring that back once it is working. Currently getting a message "I2C: not owner of 0x1201 for addr 0x12". Trying to close in on where that is coming from. Temporarily going to only read from odom to see if that is working. The flow values themselves looked ok.