Addressing the previously discussed issue by removing erroneous semaphore releases
The code was an absolute mess at this stage, and this was shortly after removing the "one dataset read per callback" approach. As a result, there was a semaphore release in between the writing of the command and the reading into the buffer. The interesting thing is that it seemed to work for odometry despite this which shouldn't happen. It did break as expected for the objdet reading. This next test will see if odometry continues to successfully read after the removal of the semaphore release. The next step after that will be bringing back objdet reading.