Attempting to read objdet datasets sequentially after odometry
The last test was successful and reading odometry was fine after removing that semaphore release. Hopefully with the semaphore release statements removed from the objdet datasets as well they will read in. This first test will not validate the data itself but will test to see whether something breaks and the odometry data stops updating.