Attempting to read odom twice
As a new approach for identifying the issue, I am going to try and call the function to read odometry twice in a single get_data() call. My hope is that this will provide insight into whether there is an issue with requesting datasets back to back or an issue with the objdet datasets.