Added a delay in between reads
Some interesting behavior. The test file is printing the flow values, not the accumulations and the implementation in this code is to actually calculate the dt value, which I think was causing it to divide by 0 and result in NaNs. A quick solution here is to add a 5 ms delay between calls, a better solution is no delay and edit the test file to print accumulations