Attempting to gain insight into where reading objdet is breaking
It doesn't seem to be an issue of not giving the sensor enough time, as consecutive reads of odometry with no delay worked fine. The next step is to print within write bytes. I suspect that maybe read objdet is not working at all. I should see it fail to write bytes immediately.